# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Director API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}": {
      "get": {
        "summary": "Get Project",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Project V1 Projects  Project Id  Get"
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema.",
                  "properties": {
                    "type": {
                      "description": "Stable problem type identifier (urn:avd:problem:...)",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "description": "Short, human-readable summary",
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "description": "HTTP status code",
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "description": "Human-readable explanation specific to this occurrence",
                      "title": "Detail",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema.",
                  "properties": {
                    "type": {
                      "description": "Stable problem type identifier (urn:avd:problem:...)",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "description": "Short, human-readable summary",
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "description": "HTTP status code",
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "description": "Human-readable explanation specific to this occurrence",
                      "title": "Detail",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema.",
                  "properties": {
                    "type": {
                      "description": "Stable problem type identifier (urn:avd:problem:...)",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "description": "Short, human-readable summary",
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "description": "HTTP status code",
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "description": "Human-readable explanation specific to this occurrence",
                      "title": "Detail",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema.",
                  "properties": {
                    "type": {
                      "description": "Stable problem type identifier (urn:avd:problem:...)",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "description": "Short, human-readable summary",
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "description": "HTTP status code",
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "description": "Human-readable explanation specific to this occurrence",
                      "title": "Detail",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects": {
      "post": {
        "summary": "Create Project",
        "operationId": "create_project_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Create Project V1 Projects Post"
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type",
                      "description": "Stable problem type identifier (urn:avd:problem:...)"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status",
                      "description": "HTTP status code"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail",
                      "description": "Human-readable explanation specific to this occurrence"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type",
                      "description": "Stable problem type identifier (urn:avd:problem:...)"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status",
                      "description": "HTTP status code"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail",
                      "description": "Human-readable explanation specific to this occurrence"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type",
                      "description": "Stable problem type identifier (urn:avd:problem:...)"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status",
                      "description": "HTTP status code"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail",
                      "description": "Human-readable explanation specific to this occurrence"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type",
                      "description": "Stable problem type identifier (urn:avd:problem:...)"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status",
                      "description": "HTTP status code"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail",
                      "description": "Human-readable explanation specific to this occurrence"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/organizations": {
      "post": {
        "summary": "Create Organization",
        "description": "Bootstrap path: create an organization and make the caller a member.\n\nThe caller must not already belong to an organization (by claim or by\nmembership), or the request is rejected (an identity belongs to one org at\na time in Phase 1).",
        "operationId": "create_organization_v1_organizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Create Organization V1 Organizations Post"
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type",
                      "description": "Stable problem type identifier (urn:avd:problem:...)"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status",
                      "description": "HTTP status code"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail",
                      "description": "Human-readable explanation specific to this occurrence"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type",
                      "description": "Stable problem type identifier (urn:avd:problem:...)"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status",
                      "description": "HTTP status code"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail",
                      "description": "Human-readable explanation specific to this occurrence"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "RFC 9457 problem-details error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type",
                      "description": "Stable problem type identifier (urn:avd:problem:...)"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title",
                      "description": "Short, human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status",
                      "description": "HTTP status code"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail",
                      "description": "Human-readable explanation specific to this occurrence"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "Problem",
                  "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrganizationRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateOrganizationRequest"
      },
...
      "CreateProjectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Stable problem type identifier (urn:avd:problem:...)"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short, human-readable summary"
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation specific to this occurrence"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem-details envelope, declared in the OpenAPI schema."
      }
    }
  }
}